Remove detailed view in discovery dashboard
