+++ packages/backend/src/modules/update-monitor/api/UpdateMonitorRouter.ts
@@ -1,7 +1,5 @@
 import Router from '@koa/router'
-import { z } from 'zod'
 
-import { withTypedContext } from '../../../api/types'
 import type { UpdateMonitorController } from './UpdateMonitorController'
 
 export function createUpdateMonitorRouter(
@@ -12,24 +10,6 @@ export function createUpdateMonitorRouter(
   router.get('/status/discovery', async (ctx) => {
     ctx.body = await updateMonitorController.getDiscoveryDashboard()
   })
-
-  router.get(
-    '/status/discovery/:chain/:project',
-    withTypedContext(
-      z.object({
-        params: z.object({
-          chain: z.string(),
-          project: z.string(),
-        }),
-      }),
-      async (ctx) => {
-        ctx.body = await updateMonitorController.getDiscoveryDashboardProject(
-          ctx.params.project,
-          ctx.params.chain,
-        )
-      },
-    ),
-  )
 
   router.get('/discovery/changes', async (ctx) => {
     ctx.set('Access-Control-Allow-Origin', '*')
