--- packages/backend/src/modules/update-monitor/api/props/dashboardContracts/abiToArray.ts
@@ -1,5 +1,0 @@
-import type { ethers } from 'ethers'
-
-export function abiToArray(abi: ethers.utils.Interface): string[] {
-  return Object.entries(abi.functions).map(([name]) => name)
-}
