--- packages/backend/src/modules/update-monitor/api/props/dashboardContracts/getFieldName.ts
@@ -1,18 +1,0 @@
-import type { ethers } from 'ethers'
-
-import { abiToArray } from './abiToArray'
-
-export function getFieldName(
-  viewABI: ethers.utils.Interface,
-  field: string,
-): string {
-  const correspondingViewFunction = abiToArray(viewABI).find(
-    (fn) => fn.split('(')[0] === field,
-  )
-
-  if (correspondingViewFunction) {
-    return correspondingViewFunction
-  }
-
-  return field
-}
