--- packages/backend/src/modules/update-monitor/api/view/constants.ts
@@ -1,13 +1,0 @@
-export const DASHBOARD_COLORS = {
-  WATCHED: '#52BE80',
-  IGNORED_IN_WATCH_MODE: '#F39C12',
-  IGNORED: '#C0392B',
-  NOT_HANDLED: '#5D6D7E',
-  INITIAL: '#F7DC6F',
-  UNVERIFIED: '#F50404',
-  DISCOVERED: '#5DADE2',
-  PROXY: '#f7237c',
-  PROXY_BACKGROUND: '#8f0d43',
-  VALUE: '#939292',
-  COMMENT: '#CDCDCD',
-}
